+++ src/core/components.ts
@@ -6,6 +6,7 @@ import * as dropdownModule from '../components/dropdown/dropdownDirective';
 import * as iconModule from '../components/icon/iconDirective';
 import * as linkModule from '../components/link/linkDirective';
 import * as overlayModule from '../components/overlay/overlayDirective';
+import * as progressIndicatorModule from '../components/progressindicator/progressIndicatorDirective';
 import * as searchboxModule from '../components/searchbox/searchboxDirective';
 import * as spinnerModule from '../components/spinner/spinnerDirective';
 import * as tableModule from '../components/table/tableDirective';
@@ -27,6 +28,7 @@ export var module: ng.IModule = ng.module('officeuifabric.components', [
   iconModule.module.name,
   linkModule.module.name,
   overlayModule.module.name,
+  progressIndicatorModule.module.name,
   searchboxModule.module.name,
   spinnerModule.module.name,
   tableModule.module.name,